datalake2/requirements.txt
pandas
boto3
mysql-connector-python
pymongo
fastapi
uvicorn
apache-airflow
pymysql

